requirements.txt
# Vouch Application Requirements
# Compatible with Python 3.11+

# Web Framework
fastapi>=0.110.0
uvicorn[standard]>=0.28.0
python-multipart>=0.0.9
jinja2>=3.1.3

# Database & Search
motor>=3.4.0
pymongo>=4.6.2
elasticsearch[async]==8.12.0

# LLM Providers
httpx>=0.27.0              # Ollama API client
openai>=1.10.0             # OpenAI API client
google-genai>=0.1.0        # Google Gemini API client (new SDK)

# Configuration & Validation
python-dotenv>=1.0.1
pydantic>=2.6.0
pydantic-settings>=2.2.1
jsonschema>=4.21.1

# File Processing
Pillow>=10.2.0
pypdf>=4.0.0
pdf2image>=1.17.0

# Development Tools
black>=26.1.0
isort>=5.13.0
flake8>=7.0.0
bandit>=1.7.5
pre-commit>=3.6.0
pytest>=9.0.2
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
